Revert "feat: 10 sprint"
