+++ src/main/java/ru/yandex/practicum/filmorate/controller/FilmController.java
@@ -1,60 +1,63 @@
 package ru.yandex.practicum.filmorate.controller;
 
-import org.springframework.beans.factory.annotation.Autowired;
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
 import org.springframework.web.bind.annotation.*;
+import ru.yandex.practicum.filmorate.exceptions.FilmorateException;
 import ru.yandex.practicum.filmorate.model.Film;
-import ru.yandex.practicum.filmorate.service.FilmService;
-import ru.yandex.practicum.filmorate.storage.film.FilmStorage;
-import ru.yandex.practicum.filmorate.storage.film.InMemoryFilmStorage;
 
 import javax.validation.Valid;
+import java.time.LocalDate;
+import java.time.Month;
+import java.util.ArrayList;
+import java.util.HashMap;
 import java.util.List;
+import java.util.Map;
 
 @RestController
 public class FilmController {
-
-    private final FilmService filmService;
-    private final FilmStorage filmStorage;
-
-    @Autowired
-    public FilmController(InMemoryFilmStorage filmStorage) {
-        this.filmStorage = filmStorage;
-        this.filmService = new FilmService(filmStorage);
-    }
+    private static final LocalDate minDate = LocalDate.of(1895, Month.DECEMBER, 28);
+    private static final Logger log = LoggerFactory.getLogger(FilmController.class);
+    private Map<Integer, Film> films = new HashMap<>();
+    private int id = 0;
 
     @PostMapping("/films")
     public Film addFilm(@Valid @RequestBody Film film) {
-        return filmStorage.addFilm(film);
+        if (checkFilm(film)) {
+            film.setId(getNextId());
+            log.debug(film.toString());
+            films.put(film.getId(), film);
+        }
+        return film;
     }
 
     @GetMapping("/films")
     public List<Film> getFilms() {
-        return filmStorage.getAllFilms();
-    }
-
-    @GetMapping("/films/{id}")
-    public Film getFilm(@PathVariable Long id) {
-        return filmStorage.getFilm(id);
-    }
-
-    @GetMapping("/films/popular")
-    public List<Film> getPopular(@RequestParam(defaultValue = "10") int count) {
-        return filmService.getPopular(count);
+        return new ArrayList<>(films.values());
     }
 
     @PutMapping("/films")
     public Film updateFilm(@RequestBody Film film) {
-        return filmStorage.updateFilm(film);
+        if (films.containsKey(film.getId())) {
+            log.debug(film.toString());
+            films.put(film.getId(), film);
+        } else throw new FilmorateException("Нет такого id");
+        return film;
     }
 
-    @PutMapping("/films/{id}/like/{userId}")
-    public void addLike(@PathVariable Long id, @PathVariable Long userId) {
-        filmService.addLike(id, userId);
+    private boolean checkFilm(Film film) {
+        if (film.getName().isBlank()) {
+            throw new FilmorateException("Имя фильма не должно быть пустым");
+        } else if (film.getDescription().length() > 200) {
+            throw new FilmorateException("Описание фильма не должно быть больше 200 символов");
+        } else if (film.getReleaseDate().isBefore(minDate)) {
+            throw new FilmorateException("Дата фильма не может быть раньше 28.12.1985");
+        } else if (film.getDuration() <= 0) {
+            throw new FilmorateException("Длительность фильма должна быть положительной");
+        } else return true;
     }
 
-    @DeleteMapping("/films/{id}/like/{userId}")
-    public void deleteLike(@PathVariable Long id, @PathVariable Long userId) {
-        filmService.deleteLike(id, userId);
+    private Integer getNextId() {
+        return ++id;
     }
-
 }
