+++ src/main/java/ru/yandex/practicum/filmorate/exceptions/FilmorateException.java
@@ -1,0 +1,8 @@
+package ru.yandex.practicum.filmorate.exceptions;
+
+public class FilmorateException extends RuntimeException {
+
+    public FilmorateException(String message) {
+        super(message);
+    }
+}
