--- src/main/java/ru/yandex/practicum/filmorate/exceptions/UserNotFoundException.java
@@ -1,13 +1,0 @@
-package ru.yandex.practicum.filmorate.exceptions;
-
-public class UserNotFoundException extends RuntimeException {
-    private final Long parameter;
-
-    public UserNotFoundException(Long parameter) {
-        this.parameter = parameter;
-    }
-
-    public String getParameter() {
-        return parameter.toString();
-    }
-}
