--- src/main/java/ru/yandex/practicum/filmorate/exceptions/ValidationException.java
@@ -1,8 +1,0 @@
-package ru.yandex.practicum.filmorate.exceptions;
-
-public class ValidationException extends RuntimeException {
-
-    public ValidationException(String message) {
-        super(message);
-    }
-}
