+++ src/test/java/ru/yandex/practicum/filmorate/controller/UserControllerTest.java
@@ -2,9 +2,8 @@ package ru.yandex.practicum.filmorate.controller;
 
 import org.junit.jupiter.api.BeforeAll;
 import org.junit.jupiter.api.Test;
-import ru.yandex.practicum.filmorate.exceptions.ValidationException;
+import ru.yandex.practicum.filmorate.exceptions.FilmorateException;
 import ru.yandex.practicum.filmorate.model.User;
-import ru.yandex.practicum.filmorate.storage.user.InMemoryUserStorage;
 
 import java.time.LocalDate;
 import java.time.Month;
@@ -14,7 +13,6 @@ import static org.junit.jupiter.api.Assertions.assertThrows;
 
 class UserControllerTest {
     private static UserController userC;
-    private static final InMemoryUserStorage userStorage = new InMemoryUserStorage();
     private static User user1;
     private static User user2;
     private static User user3;
@@ -26,7 +24,7 @@ class UserControllerTest {
         user2 = new User("2", "Login1", "123", LocalDate.of(1994, Month.SEPTEMBER, 9));
         user3 = new User("3", "Login1", "123@", LocalDate.of(2994, Month.SEPTEMBER, 9));
         user4 = new User("Nick Name", "dolore", "mail@mail.ru", LocalDate.of(1946, Month.AUGUST, 20));
-        userC = new UserController(userStorage);
+        userC = new UserController();
     }
 
     @Test
@@ -34,11 +32,11 @@ class UserControllerTest {
         assertEquals(user1, userC.addUser(user1));
         assertEquals(user4, userC.addUser(user4));
         assertThrows(
-                ValidationException.class,
+                FilmorateException.class,
                 () -> userC.addUser(user2)
         );
         assertThrows(
-                ValidationException.class,
+                FilmorateException.class,
                 () -> userC.addUser(user3)
         );
     }
